pyproject.toml
[project]
name = "agentic-ai-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.6",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
    "tex>=1.8",
]
